Reorg - get rid of cli Util object
